+++ example/lib/full_map.dart
@@ -36,7 +36,7 @@ class FullMapState extends State<FullMap> {
   }
 
   _onCameraChangeListener(CameraChangedEventData data) {
-    print("CameraChangedEventData: timestamp: ${data.timestamp}");
+    print("CameraChangedEventData: ${data.debugInfo}");
   }
 
   _onResourceRequestListener(ResourceEventData data) {
@@ -140,5 +140,19 @@ class FullMapState extends State<FullMap> {
           onResourceRequestListener: _onResourceRequestListener,
           onLongTapListener: (coordinate) {},
         ));
+  }
+}
+
+extension on CameraChangedEventData {
+
+  String get debugInfo {
+    return "timestamp ${DateTime.fromMicrosecondsSinceEpoch(timestamp)}, camera: ${cameraState.debugInfo}";
+  }
+}
+
+extension on CameraState {
+
+  String get debugInfo {
+    return "lat: ${center.coordinates.lat}, lng: ${center.coordinates.lng}, zoom: ${zoom}, bearing: ${bearing}, pitch: ${pitch}";
   }
 }