+++ ios/Classes/Extensions.swift
@@ -5,7 +5,7 @@ import Flutter
 
 let COORDINATES = "coordinates"
 
-extension FlutterError: Error { }
+extension FlutterError: @retroactive Error { }
 
 // FLT to Mapbox
 
@@ -610,6 +610,7 @@ func convertDictionaryToString(dict: [String?: Any?]?) -> String {
 }
 
 func convertDictionaryToGeometry(dict: [String?: Any?]?) -> Geometry? {
+    guard let dict else { return nil }
     do {
         let jsonData = try JSONSerialization.data(withJSONObject: dict, options: JSONSerialization.WritingOptions.init(rawValue: 0))
         let geometry = try JSONDecoder().decode(Geometry.self, from: jsonData)
@@ -1033,6 +1034,33 @@ extension MapboxCommon.NetworkRestriction {
         }
     }
 }
+
+extension _TileStoreOptionsKey {
+
+    func toTileStoreOptionsKey() -> String {
+        switch self {
+        case .dISKQUOTA:
+            return TileStoreOptions.diskQuota
+        case .mAPBOXAPIURL:
+            return TileStoreOptions.mapboxAPIURL
+        case .tILEURLTEMPLATE:
+            return TileStoreOptions.tileURLTemplate
+        }
+    }
+}
+
+extension TileDataDomain {
+
+    func toTileDataDomain() -> MapboxCommon.TileDataDomain {
+        switch self {
+        case .mAPS: return .maps
+        case .nAVIGATION: return .navigation
+        case .sEARCH: return .search
+        case .aDAS: return .adas
+        }
+    }
+}
+
 // MARK: Result
 extension Result where Failure == any Error {
     init(code: String, catchingFlutter body: () throws -> Success) {